Update scene_major_one_minor_one_prestages.js_definition
This seemed to be the only way networked data was allowed to be loaded into the OpenGL hardware when using more than one file of differing response types, which was creating multiple xmlhttprequest objects instead of reusing the same one. Which by c and c++ standards(which most if not all browsers are derived from) should have operated as such.